client: fix implementation to work against rust language server
This implements a bunch more features for our initial project. It allows
the rust language server to communicate with jsonrpc-glib. In particular,
we had to switch to "id" fields of "int" rather than string. This limits
the longevity of our client, but should be fine.

Additionally JsonrpcClient was improved to perform async read loops
to deliver notifications out-of-band. This is useful so the consumer
of JsonrpcClient needs to only connect to JsonrpcClient::notification
and filter for their ideal notification. I considered using detailed
quarks for these messages, but that could leak random strings into
our quark table.

I've also added some rudimentary documentation, but we still need a
reference manual so that won't directly help consumers yet.

This also adds the "Content-Length" header to the requests, which I'm
not sure is a standardized thing or not. I haven't seen it anywhere
on the JSON-RPC spec.

We might be able to change things to allow ourselves to deal with
implementation details like 'Content-Length' by using subclasses
and give ourselves a "discover message length" hook.

I'd be interested in people testing this against other JSON-RPC
servers to see what breaks. Certainly there will be things to
improve.