jcon: add JCON_EXTRACT()
This allows you to extract items from a JsonNode with a syntax similar to
that of JCON_NEW().

For example, you can for the most part just change "JCON_*" to "JCONE_" in
the syntax to extract from a node (in addition to adding the JsonNode
as the first parameter).

gboolean success = JCON_EXTRACT (
  "foo", JCONE_STRING (foo)
);