gvariant: use GVariant based public API
By switching to a GVariant based public API (similar to say, GDBus), we
allow ourselves to implement an opitmization in the network layer. For
example, when talking between two peers that are jsonrpc-glib based, we
can upgrade the connection to application/gvariant instead of
application/json. In doing so, we reduce our overhead for parsing peer
content, allow ourselves to avoid parse checks in trusted situations, and
lower the allocator overhead for small objects used by JsonNode trees.

This means that when recieving large replies, consumers can lower memory
fragmentation by using GVariant from point-to-point and lazily access the
GVariant contents.