server: add handler convenience API
Rather than requiring that all handlers be dispatched via connecting to
the handle-call signal, this allows a simplier API to register handlers
and remove them as necessary.

To keep things simple, we just use a sorted GArray of handler data and
bsearch to locate the proper handler upon dispatch. For jsonrpc, this
is likely better than a tree of nodes that we parse parameters from
because methods generally don't have parseable user data like a HTTP url
might have.