message: fix va_list use on architectures with compound structs
Not all architectures use the design that x86_64 uses with va_list where it
is a pointer. Some architectures (such as aarch64) use a compound struct
and therefore it is not suitable for passing as an argument.

This ensures that we pass things as a va_list* so that the original
va_list location state is incremented to walk through the register/stack
items.